Rudy's data challenge's result
I didn't meet any difficulty for the SQL piece of work. Regarding the cleansing, buiding and the modeling, I'm not used to do that in Python and I don't have enough time to do it well so I simply wrote some sudo code to give an idea of my understanding of the problem and how I think we should fix it.
I'm happy to discuss that further with you.